Add Sparky class
The content lengths match, so I'm going to cautiously proceed under the
assumption that the output is the same across the board. There must be a better
way of getting this code into a testable state. The book skips testability for
this chapter, but I didn't want to. Realistically, it's the first thing that
must be done. _Working Effectively with Legacy Code_ has some content around
getting untested classes into a test harness. However, in this case, we don't
even have a class. We have a bunch of script code dumped into a global context.
It occurs to me that simply assigning the various strings to variables and
comparing them may be good enough. I'll try that next.